builder: Move address builder tests from tx to address.
