+++ core/src/builder/transaction.rs
@@ -342,127 +342,9 @@ pub fn create_tx_outs(pairs: Vec<(Amount, ScriptBuf)>) -> Vec<TxOut> {
 
 #[cfg(test)]
 mod tests {
-    use crate::{
-        builder,
-        musig2::AggregateFromPublicKeys,
-        utils::{self, SECP},
-    };
-    use bitcoin::{
-        hashes::Hash, key::TapTweak, Address, AddressType, OutPoint, ScriptBuf, Txid,
-        XOnlyPublicKey,
-    };
-    use secp256k1::{rand, Keypair, PublicKey, SecretKey};
-    use std::str::FromStr;
-
-    #[test]
-    fn create_taproot_address() {
-        let secret_key = SecretKey::new(&mut rand::thread_rng());
-        let internal_key =
-            XOnlyPublicKey::from_keypair(&Keypair::from_secret_key(&SECP, &secret_key)).0;
-
-        // No internal key or scripts (key path spend).
-        let (address, spend_info) =
-            builder::address::create_taproot_address(&[], None, bitcoin::Network::Regtest);
-        assert_eq!(address.address_type().unwrap(), AddressType::P2tr);
-        assert!(address.is_related_to_xonly_pubkey(
-            &utils::UNSPENDABLE_XONLY_PUBKEY
-                .tap_tweak(&SECP, spend_info.merkle_root())
-                .0
-                .to_inner()
-        ));
-        assert_eq!(spend_info.internal_key(), *utils::UNSPENDABLE_XONLY_PUBKEY);
-        assert!(spend_info.merkle_root().is_none());
-
-        // Key path spend.
-        let (address, spend_info) = builder::address::create_taproot_address(
-            &[],
-            Some(internal_key),
-            bitcoin::Network::Regtest,
-        );
-        assert_eq!(address.address_type().unwrap(), AddressType::P2tr);
-        assert!(address.is_related_to_xonly_pubkey(
-            &internal_key
-                .tap_tweak(&SECP, spend_info.merkle_root())
-                .0
-                .to_inner()
-        ));
-        assert_eq!(spend_info.internal_key(), internal_key);
-        assert!(spend_info.merkle_root().is_none());
-
-        let scripts = [ScriptBuf::new()];
-        let (address, spend_info) = builder::address::create_taproot_address(
-            &scripts,
-            Some(internal_key),
-            bitcoin::Network::Regtest,
-        );
-        assert_eq!(address.address_type().unwrap(), AddressType::P2tr);
-        assert!(address.is_related_to_xonly_pubkey(
-            &internal_key
-                .tap_tweak(&SECP, spend_info.merkle_root())
-                .0
-                .to_inner()
-        ));
-        assert_eq!(spend_info.internal_key(), internal_key);
-        assert!(spend_info.merkle_root().is_some());
-
-        let scripts = [ScriptBuf::new(), ScriptBuf::new()];
-        let (address, spend_info) = builder::address::create_taproot_address(
-            &scripts,
-            Some(internal_key),
-            bitcoin::Network::Regtest,
-        );
-        assert_eq!(address.address_type().unwrap(), AddressType::P2tr);
-        assert!(address.is_related_to_xonly_pubkey(
-            &internal_key
-                .tap_tweak(&SECP, spend_info.merkle_root())
-                .0
-                .to_inner()
-        ));
-        assert_eq!(spend_info.internal_key(), internal_key);
-        assert!(spend_info.merkle_root().is_some());
-    }
-
-    #[test]
-    fn generate_deposit_address_musig2_fixed_address() {
-        let verifier_pks_hex: Vec<&str> = vec![
-            "034f355bdcb7cc0af728ef3cceb9615d90684bb5b2ca5f859ab0f0b704075871aa",
-            "02466d7fcae563e5cb09a0d1870bb580344804617879a14949cf22285f1bae3f27",
-            "023c72addb4fdf09af94f0c94d7fe92a386a7e70cf8a1d85916386bb2535c7b1b1",
-            "032c0b7cf95324a07d05398b240174dc0c2be444d96b159aa6c7f7b1e668680991",
-            "029ac20335eb38768d2052be1dbbc3c8f6178407458e51e6b4ad22f1d91758895b",
-            "035ab4689e400a4a160cf01cd44730845a54768df8547dcdf073d964f109f18c30",
-            "037962d45b38e8bcf82fa8efa8432a01f20c9a53e24c7d3f11df197cb8e70926da",
-        ];
-        let verifier_pks: Vec<PublicKey> = verifier_pks_hex
-            .iter()
-            .map(|pk| PublicKey::from_str(pk).unwrap())
-            .collect();
-        let nofn_xonly_pk = XOnlyPublicKey::from_musig2_pks(verifier_pks, None, false);
-
-        let evm_address: [u8; 20] = hex::decode("1234567890123456789012345678901234567890")
-            .unwrap()
-            .try_into()
-            .unwrap();
-
-        let recovery_taproot_address =
-            Address::from_str("bcrt1p65yp9q9fxtf7dyvthyrx26xxm2czanvrnh9rtvphmlsjvhdt4k6qw4pkss")
-                .unwrap();
-
-        let deposit_address = builder::address::generate_deposit_address(
-            nofn_xonly_pk,
-            recovery_taproot_address.as_unchecked(),
-            crate::EVMAddress(evm_address),
-            100_000_000,
-            bitcoin::Network::Regtest,
-            200,
-        );
-
-        // Comparing it to the taproot address generated in bridge backend.
-        assert_eq!(
-            deposit_address.0.to_string(),
-            "bcrt1ptlz698wumzl7uyk6pgrvsx5ep29thtvngxftywnd4mwq24fuwkwsxasqf5" // TODO: check this later
-        )
-    }
+    use crate::{builder, utils::SECP};
+    use bitcoin::{hashes::Hash, OutPoint, Txid, XOnlyPublicKey};
+    use secp256k1::{rand, Keypair, SecretKey};
 
     #[test]
     fn create_move_tx() {
